infra/lambda/summary_news/pyproject.toml
[tool.poetry]
name = "summary-news"
version = "0.1.0"
description = ""
authors = ["yoshi-d-24"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.3.1"
ipython = "^8.18.1"
boto3 = "^1.34.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
